devices/requirements.txt
counterfit==0.1.4.dev9
counterfit-connection==0.1.0.dev5
counterfit-shims-grove==0.1.4.dev5
werkzeug==2.3.8
azure-core==1.30.2
azure-iot-device==2.14.0
azure-iot-hub==2.6.1